requirements.txt
beautifulsoup4
coloredlogs
Flask
gunicorn
python-dateutil
python-dotenv
requests
